+++ mw/index.ts
@@ -9,7 +9,7 @@ declare const $STYLE: string;
 
 // 每次新增插件都需要修改这里
 const baseVersion = '3.16',
-	addons = ['blockCursor'];
+	addons = ['blockCursor', 'indentGuide'];
 
 mw.loader.addStyleTag($STYLE);